SANKOFA 0/requirements.txt
Django==5.2.8
Pillow==12.0.0
django-ckeditor==6.7.3
django-allauth
requests
PyJWT
cryptography
